requirements.txt
certifi==2018.1.18
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11
django-crispy-forms==1.7.2
django-filter==1.1.0
django-formtools==2.1
django-guardian==1.4.9
djangorestframework==3.6.3
et-xmlfile==1.0.1
future==0.16.0
httplib2==0.11.3
idna==2.6
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
mysqlclient==1.3.12
odfpy==1.3.6
openpyxl==2.5.2
Pillow==5.1.0
pytz==2018.4
requests==2.18.4
six==1.11.0
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
wincertstore==0.2
xlrd==1.1.0
XlsxWriter==1.0.4
xlwt==1.3.0

